+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 - Added option to enter reader mode when tapping on a link in iOS
 - Support new scaled and controversial sort types - contribution from @micahmo
 - Added support to open Lemmy links in app. Android only.  - contribution from @ggichure
+- Added support for display cross-posts - contribution from @micahmo
 
 ### Changed
 - Collapsed comments are easier to expand - contribution from @micahmo